requirements.txt
crewai
langchain-groq
langchain-community
duckduckgo-search
python-dotenv